refactor: CustomAppBar 섹션별 위젯 분리 #21
