+++ lib/common/index.dart
@@ -1,2 +1,2 @@
 export 'package:typed/common/layout/default_layout.dart';
-export 'package:typed/common/widgets/custom_app_bar.dart';
+export 'package:typed/common/widgets/app_bar/custom_app_bar.dart';
